#624 fix: canonical tag bug causing GSC indexing failure
Root layout's default alternates.canonical: '/' was being inherited by
/privacy, /terms, and /leaderboard, contradicting the sitemap and
likely causing Google to fold these pages into the homepage instead
of indexing them. /leaderboard couldn't export metadata at all since
it was a 'use client' page; split into server page.tsx + new
LeaderboardClient.tsx. Added generateMetadata to /u/[publicProfileId]
(respects profileVisibility). Also fixed duplicate "| Boardly | Boardly"
title suffix on privacy, terms, leaderboard, and games pages.
